my-portfolio/requirements.txt
click==8.1.3
Flask==2.2.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
python-decouple==3.6
Werkzeug==2.2.2
